essentia-tensorflow/requirements.txt
essentia-tensorflow